requirements/staging.txt
dj-database-url==0.3.0
django-toolbelt==0.0.1
cloudinary==1.2.0